#35 重构VJMap代码
